Return null if given extra characters
QMJsonValue::fromJson() would return valid QMJsonValues if the
provided string had extra non-whitespace characters provided.
Although this is likely acceptable with respect to the JSON
spec, it should be treated as an error.

The following adds a check to the end of the parse function
to validate that no extra non-whitespace characters were
added to the JSON string.

[ISSUE]: https://github.com/QtMark/qmjson/issues/9

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>